services/data_adapter_service/requirements.txt
aiopulsar-py==0.1.0a0
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
astunparse==1.6.3
attrs==22.1.0
botocore==1.29.114
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==6.5.0
cryptography==42.0.5
distlib==0.3.6
exceptiongroup==1.2.0
Faker==14.0.0
fastapi==0.110.0
fastavro==1.9.4
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT==0.2.0
Flask-JWT-Extended==4.2.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
git-remote-codecommit==1.16
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
libcst==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.12.1
marshmallow-sqlalchemy==0.25.0
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==21.3
paho-mqtt==1.6.1
platformdirs==2.6.2
pluggy==1.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pulsar-client==3.4.0
py==1.11.0
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyDispatcher==2.0.6
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.28.2
rich==13.7.1
sendgrid==6.9.7
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.44
starkbank-ecdsa==2.2.0
starlette==0.36.3
strawberry-graphql==0.220.0
tomli==2.0.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==1.26.15
uvicorn==0.28.0
virtualenv==20.17.1
websockets==12.0
Werkzeug==2.2.2
zipp==3.17.0
